minor restructuring/renaming
