+++ packages/api/src/city/city-exists.constraint.ts
@@ -1,14 +1,14 @@
 import { Injectable } from '@nestjs/common';
 import { ValidatorConstraint, ValidatorConstraintInterface } from 'class-validator';
-import { ChainableCityProvider } from '../interfaces/chainable-city.provider';
+import { ChainableCityProvider } from './interfaces/chainable-city.provider';
 
 @ValidatorConstraint({ async: true })
 @Injectable()
 export class CityExistsConstraint implements ValidatorConstraintInterface {
-  constructor(private readonly chain: ChainableCityProvider) {}
+  constructor(private readonly cityProvider: ChainableCityProvider) {}
 
   async validate(value: string) {
-    return await this.chain.handle(value);
+    return await this.cityProvider.handle(value);
   }
 
   defaultMessage() {
