+++ packages/api/src/city/city.module.ts
@@ -1,14 +1,13 @@
-import { HttpModule } from '@nestjs/axios';
 import { Module } from '@nestjs/common';
 import { LoggerModule } from 'src/logger/logger.module';
-import { CityExistsConstraint } from './constraints/city-exists.constraint';
+import { CityExistsConstraint } from './city-exists.constraint';
 import { CityProviderFactory } from './factories/city-provider.factory';
 import { ChainableCityProvider } from './interfaces/chainable-city.provider';
 import { OpenWeatherCityProvider } from './providers/openweather.provider';
 import { WeatherApiCityProvider } from './providers/weatherapi.provider';
 
 @Module({
-  imports: [HttpModule.register({}), LoggerModule],
+  imports: [LoggerModule],
   providers: [
     CityExistsConstraint,
     WeatherApiCityProvider,
