+++ packages/api/src/weather/weather.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
-import { CurrentWeatherResponseDto } from '../dtos/current-weather-response.dto';
-import { ChainableCurrentWeatherProvider } from '../interfaces/chainable-weather-provider.abstract';
-import { WeatherProvider } from '../interfaces/current-weather.abstract';
+import { CurrentWeatherResponseDto } from './dtos/current-weather-response.dto';
+import { ChainableCurrentWeatherProvider } from './interfaces/chainable-weather-provider.abstract';
+import { WeatherProvider } from './interfaces/current-weather.abstract';
 
 @Injectable()
 export class WeatherService implements WeatherProvider {
